LayoutTests/imported/w3c/web-platform-tests/streams/piping/pipe-through.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: undefined

FAIL Piping through a duck-typed pass-through transform stream should work |this| is not a Promise
FAIL Piping through a transform errored on the writable end does not cause an unhandled promise rejection |this| is not a Promise
FAIL pipeThrough should not call pipeTo on this |this| is not a Promise
FAIL pipeThrough should not call pipeTo on the ReadableStream prototype assert_false: the monkey-patched pipeTo should not have been called expected false got true
PASS pipeThrough should brand-check this and not allow 'null' 
FAIL pipeThrough should brand-check readable and not allow 'null' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check this and not allow 'undefined' 
FAIL pipeThrough should brand-check readable and not allow 'undefined' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check this and not allow '0' 
FAIL pipeThrough should brand-check readable and not allow '0' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check this and not allow 'NaN' 
FAIL pipeThrough should brand-check readable and not allow 'NaN' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check this and not allow 'true' 
FAIL pipeThrough should brand-check readable and not allow 'true' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check this and not allow 'ReadableStream' 
FAIL pipeThrough should brand-check readable and not allow 'ReadableStream' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check this and not allow '[object ReadableStream]' 
FAIL pipeThrough should brand-check readable and not allow '[object ReadableStream]' assert_false: writable should not have been accessed expected false got true
PASS pipeThrough should brand-check writable and not allow 'null' 
PASS pipeThrough should brand-check writable and not allow 'undefined' 
PASS pipeThrough should brand-check writable and not allow '0' 
PASS pipeThrough should brand-check writable and not allow 'NaN' 
PASS pipeThrough should brand-check writable and not allow 'true' 
PASS pipeThrough should brand-check writable and not allow 'WritableStream' 
PASS pipeThrough should brand-check writable and not allow '[object WritableStream]' 
PASS pipeThrough should rethrow errors from accessing readable or writable 
PASS invalid values of signal should throw; specifically 'null' 
PASS invalid values of signal should throw; specifically '0' 
PASS invalid values of signal should throw; specifically 'NaN' 
PASS invalid values of signal should throw; specifically 'true' 
PASS invalid values of signal should throw; specifically 'AbortSignal' 
PASS invalid values of signal should throw; specifically '[object AbortSignal]' 
FAIL pipeThrough should accept a real AbortSignal |this| is not a Promise
PASS pipeThrough should throw if this is locked 
PASS pipeThrough should throw if writable is locked 
FAIL pipeThrough should not care if readable is locked |this| is not a Promise
FAIL preventCancel should work |this| is not a Promise
FAIL preventClose should work |this| is not a Promise
FAIL preventAbort should work |this| is not a Promise
PASS pipeThrough() should throw if an option getter grabs a writer 
